refactor(draw): seperate into draw and edit composable
